# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recipe API",
    "version": "1.0.0",
    "description": "API pentru gestionarea rețetelor culinare"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Recipe": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "ingredients": { "type": "string" },
          "instructions": { "type": "string" },
          "prepTime": { "type": "integer" },
          "cookTime": { "type": "integer" },
          "servings": { "type": "integer" },
          "recipeType": { "type": "string" },
          "cuisine": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "image": { "type": "string" }
            }
          },
          "source": { "type": "string" },
          "userId": { "type": "integer", "nullable": true },
          "isAIGenerated": { "type": "boolean" },
          "isPrivate": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string" },
          "username": { "type": "string" },
          "role": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/users/register": {
      "post": {
        "summary": "Înregistrează un utilizator nou",
        "description": "Permite crearea unui nou utilizator. Returnează un token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "description": "Email-ul utilizatorului" },
                  "username": { "type": "string", "description": "Numele de utilizator" },
                  "password": { "type": "string", "description": "Parola utilizatorului" }
                },
                "required": ["email", "username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilizator înregistrat cu succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "description": "Token-ul JWT al utilizatorului" }
                  }
                }
              }
            }
          },
          "400": { "description": "Date incorecte sau lipsește un câmp necesar" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Autentificare utilizator",
        "description": "Permite autentificarea unui utilizator și returnează un token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "description": "Email-ul utilizatorului" },
                  "password": { "type": "string", "description": "Parola utilizatorului" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autentificare cu succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "description": "Token-ul JWT al utilizatorului" }
                  }
                }
              }
            }
          },
          "400": { "description": "Credentiale invalide" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/google-login": {
      "post": {
        "summary": "Autentificare cu Google",
        "description": "Primește tokenul de la Google și returnează un JWT pentru aplicație.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string", "description": "Tokenul ID primit de la Google" }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autentificare reușită",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Token lipsă sau invalid" },
          "500": { "description": "Eroare server" }
        }
      }
    },
...
    "/api/users/auth/google/callback": {
      "get": {
        "summary": "Callback pentru autentificare Google",
        "description": "Procesează răspunsul de la Google și autentifică utilizatorul",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": { "type": "string" },
            "description": "Codul de autorizare de la Google"
          }
        ],
        "responses": {
          "302": { "description": "Redirect către aplicație cu token JWT" },
          "400": { "description": "Cod de autorizare lipsă" },
          "500": { "description": "Eroare la autentificarea Google" }
        }
      }
    },
    "/api/users/update-username": {
      "put": {
        "summary": "Actualizează numele de utilizator",
        "description": "Actualizează numele de utilizator pentru utilizatorul autentificat",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "minLength": 3, "maxLength": 50 }
                },
                "required": ["username"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Numele de utilizator a fost actualizat cu succes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Nume de utilizator invalid sau deja existent" },
          "401": { "description": "Neautentificat" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/user-profile/{recipeId}": {
      "get": {
        "summary": "Obține profilul utilizatorului după ID-ul rețetei",
        "description": "Returnează informații despre utilizatorul care a creat o rețetă specifică",
        "parameters": [
          {
            "in": "path",
            "name": "recipeId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul rețetei"
          }
        ],
        "responses": {
          "200": {
            "description": "Informații despre utilizator",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": { "type": "integer" },
                    "username": { "type": "string" },
                    "recipeCount": { "type": "integer" },
                    "followersCount": { "type": "integer" },
                    "isSpoonacularRecipe": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "404": { "description": "Rețeta sau utilizatorul nu a fost găsit" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/recipe-author/{recipeId}": {
      "get": {
        "summary": "Obține inițialele autorului rețetei",
        "description": "Returnează inițialele utilizatorului care a creat rețeta",
        "parameters": [
          {
            "in": "path",
            "name": "recipeId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul rețetei"
          }
        ],
        "responses": {
          "200": {
            "description": "Inițialele autorului",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "initials": { "type": "string" },
                    "isUserRecipe": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "404": { "description": "Rețeta nu a fost găsită" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/delete-account": {
      "delete": {
        "summary": "Șterge contul utilizatorului",
        "description": "Șterge complet contul utilizatorului autentificat și toate datele asociate",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Contul a fost șters cu succes" },
          "401": { "description": "Neautentificat" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "summary": "Obține informații despre utilizatorul curent",
        "description": "Returnează informații despre utilizatorul autentificat",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Informații despre utilizator",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Neautentificat" },
          "404": { "description": "Utilizatorul nu a fost găsit" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/user/{userId}": {
      "get": {
        "summary": "Obține utilizatorul după ID",
        "description": "Returnează informații despre un utilizator specific",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul utilizatorului"
          }
        ],
        "responses": {
          "200": {
            "description": "Informații despre utilizator",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "username": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "_count": {
                      "type": "object",
                      "properties": {
                        "recipes": { "type": "integer" },
                        "followers": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Utilizatorul nu a fost găsit" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/recipes-by-user/{userId}": {
      "get": {
        "summary": "Obține rețetele unui utilizator cu paginare",
        "description": "Returnează rețetele create de un utilizator specific cu suport pentru paginare",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul utilizatorului"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 },
            "description": "Numărul paginii"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 12 },
            "description": "Numărul de rețete per pagină"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor utilizatorului cu informații de paginare",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recipes": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          { "$ref": "#/components/schemas/Recipe" },
                          {
                            "type": "object",
                            "properties": {
                              "user": {
                                "type": "object",
                                "properties": {
                                  "username": { "type": "string" }
                                }
                              }
                            }
                          }
                        ]
                      }
                    },
                    "totalPages": { "type": "integer" },
                    "currentPage": { "type": "integer" },
                    "totalRecipes": { "type": "integer" }
                  }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/spoonacular-recipes": {
      "get": {
        "summary": "Obține doar rețetele Spoonacular",
        "description": "Returnează rețetele fără userId (rețete Spoonacular)",
        "responses": {
          "200": {
            "description": "Lista rețetelor Spoonacular",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/followers": {
      "get": {
        "summary": "Obține urmăritorii utilizatorului curent",
        "description": "Returnează lista urmăritorilor utilizatorului autentificat",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista urmăritorilor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "username": { "type": "string" },
                      "email": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Neautentificat" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/users/following": {
      "get": {
        "summary": "Obține utilizatorii urmăriți de utilizatorul curent",
        "description": "Returnează lista utilizatorilor pe care îi urmărește utilizatorul autentificat",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista utilizatorilor urmăriți",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "username": { "type": "string" },
                      "email": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Neautentificat" },
          "500": { "description": "Eroare server" }
        }
      }
    },
...
  }
}